chore: Main 페이지 경로 변경 및 main.tsx 파일 삭제
