+++ pages/index.tsx
@@ -1,16 +1,58 @@
+import Header from '@domains/Header'
 import styled from '@emotion/styled'
+import { FONT_SIZES } from '@utils/constants/sizes'
+import { COLORS } from '@utils/constants/colors'
+import MUIButton from '@components/MUIButton'
 
-export default function Home() {
-  return <ExampleBtn>Button</ExampleBtn>
+const main = () => {
+  const onButtonClick = () => {
+    alert('안녕')
+  }
+
+  return (
+    <MainContainer>
+      <Header />
+      <Text>텍스트 컴포넌트</Text>
+      <MUIButton
+        onClick={onButtonClick}
+        style={{
+          color: 'white',
+          height: '40px',
+          width: '70%',
+          borderRadius: '50px',
+          backgroundColor: 'red',
+          position: 'absolute',
+          top: '210px',
+          zIndex: '99',
+        }}>
+        이벤트 등록하기
+      </MUIButton>
+      <VideoBox src={'/video/Christmas_Tree.mp4'} autoPlay muted />
+    </MainContainer>
+  )
 }
 
-const ExampleBtn = styled.button`
-  background-color: red;
-  width: calc(80% - 32px);
-  margin: 0 auto;
-  height: 32px;
-  border-radius: 20px;
-  display: block;
-  border: none;
-  cursor: pointer;
+const MainContainer = styled.div`
+  width: 100%;
+  height: 100%;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  position: relative;
 `
+const Text = styled.text`
+  font-size: ${FONT_SIZES.LARGE};
+  color: ${COLORS.WHITE};
+  background-color: gray;
+  position: absolute;
+  top: 120px;
+`
+
+const VideoBox = styled.video`
+  width: 100%;
+  position: absolute;
+  bottom: 0;
+  z-index: 9;
+`
+
+export default main
